+++ modules/mod_draw/mod_draw.c
@@ -29,6 +29,7 @@
 
 /* --------------------------------------------------------------------------- */
 
+#include <stdint.h>
 #include <string.h>
 
 #include <math.h>
@@ -346,14 +347,14 @@ static void _moddraw_object_move(int id, int x, int y) {
 /* --------------------------------------------------------------------------- */
 /* Exportable functions                                                        */
 
-int moddraw_drawing_map(INSTANCE *my, int *params) {
+int moddraw_drawing_map(INSTANCE *my, intptr_t *params) {
     drawing_graph = bitmap_get(params[0], params[1]);
     return 1;
 }
 
 /* --------------------------------------------------------------------------- */
 
-int moddraw_drawing_at(INSTANCE *my, int *params) {
+int moddraw_drawing_at(INSTANCE *my, intptr_t *params) {
     drawing_graph = NULL;
     drawing_z     = params[0];
     return 1;
@@ -361,42 +362,42 @@ int moddraw_drawing_at(INSTANCE *my, int *params) {
 
 /* --------------------------------------------------------------------------- */
 
-int moddraw_drawing_stipple(INSTANCE *my, int *params) {
+int moddraw_drawing_stipple(INSTANCE *my, intptr_t *params) {
     drawing_stipple = params[0];
     return 1;
 }
 
 /* --------------------------------------------------------------------------- */
 
-int moddraw_delete_drawing(INSTANCE *my, int *params) {
+int moddraw_delete_drawing(INSTANCE *my, intptr_t *params) {
     _moddraw_object_destroy(params[0]);
     return 1;
 }
 
 /* --------------------------------------------------------------------------- */
 
-int moddraw_move_drawing(INSTANCE *my, int *params) {
+int moddraw_move_drawing(INSTANCE *my, intptr_t *params) {
     _moddraw_object_move(params[0], params[1], params[2]);
     return 1;
 }
 
 /* --------------------------------------------------------------------------- */
 
-int moddraw_drawing_color(INSTANCE *my, int *params) {
+int moddraw_drawing_color(INSTANCE *my, intptr_t *params) {
     gr_setcolor(params[0]);
     return 1;
 }
 
 /* --------------------------------------------------------------------------- */
 
-int moddraw_drawing_alpha(INSTANCE *my, int *params) {
+int moddraw_drawing_alpha(INSTANCE *my, intptr_t *params) {
     gr_setalpha(params[0]);
     return 1;
 }
 
 /* --------------------------------------------------------------------------- */
 
-int moddraw_box(INSTANCE *my, int *params) {
+int moddraw_box(INSTANCE *my, intptr_t *params) {
     if (!drawing_graph) {
         DRAWING_OBJECT *dr = malloc(sizeof(DRAWING_OBJECT));
 
@@ -414,7 +415,7 @@ int moddraw_box(INSTANCE *my, int *params) {
 
 /* --------------------------------------------------------------------------- */
 
-int moddraw_rect(INSTANCE *my, int *params) {
+int moddraw_rect(INSTANCE *my, intptr_t *params) {
     if (!drawing_graph) {
         DRAWING_OBJECT *dr = malloc(sizeof(DRAWING_OBJECT));
 
@@ -433,7 +434,7 @@ int moddraw_rect(INSTANCE *my, int *params) {
 
 /* --------------------------------------------------------------------------- */
 
-int moddraw_line(INSTANCE *my, int *params) {
+int moddraw_line(INSTANCE *my, intptr_t *params) {
     if (!drawing_graph) {
         DRAWING_OBJECT *dr = malloc(sizeof(DRAWING_OBJECT));
 
@@ -451,7 +452,7 @@ int moddraw_line(INSTANCE *my, int *params) {
 
 /* --------------------------------------------------------------------------- */
 
-int moddraw_circle(INSTANCE *my, int *params) {
+int moddraw_circle(INSTANCE *my, intptr_t *params) {
     if (!drawing_graph) {
         DRAWING_OBJECT *dr = malloc(sizeof(DRAWING_OBJECT));
 
@@ -468,7 +469,7 @@ int moddraw_circle(INSTANCE *my, int *params) {
 
 /* --------------------------------------------------------------------------- */
 
-int moddraw_fcircle(INSTANCE *my, int *params) {
+int moddraw_fcircle(INSTANCE *my, intptr_t *params) {
     if (!drawing_graph) {
         DRAWING_OBJECT *dr = malloc(sizeof(DRAWING_OBJECT));
 
@@ -485,7 +486,7 @@ int moddraw_fcircle(INSTANCE *my, int *params) {
 
 /* --------------------------------------------------------------------------- */
 
-int moddraw_bezier(INSTANCE *my, int *params) {
+int moddraw_bezier(INSTANCE *my, intptr_t *params) {
     if (!drawing_graph) {
         DRAWING_OBJECT *dr = malloc(sizeof(DRAWING_OBJECT));
 
@@ -509,20 +510,20 @@ int moddraw_bezier(INSTANCE *my, int *params) {
 
 /* --------------------------------------------------------------------------- */
 
-int moddraw_get_pixel(INSTANCE *my, int *params) {
+int moddraw_get_pixel(INSTANCE *my, intptr_t *params) {
     return gr_get_pixel(background, params[0], params[1]);
 }
 
 /* --------------------------------------------------------------------------- */
 
-int moddraw_put_pixel(INSTANCE *my, int *params) {
+int moddraw_put_pixel(INSTANCE *my, intptr_t *params) {
     gr_put_pixel(background, params[0], params[1], params[2]);
     return 1;
 }
 
 /* --------------------------------------------------------------------------- */
 
-int moddraw_map_get_pixel(INSTANCE *my, int *params) {
+int moddraw_map_get_pixel(INSTANCE *my, intptr_t *params) {
     GRAPH *map = bitmap_get(params[0], params[1]);
     if (!map)
         return -1;
@@ -531,7 +532,7 @@ int moddraw_map_get_pixel(INSTANCE *my, int *params) {
 
 /* --------------------------------------------------------------------------- */
 
-int moddraw_map_put_pixel(INSTANCE *my, int *params) {
+int moddraw_map_put_pixel(INSTANCE *my, intptr_t *params) {
     GRAPH *map = bitmap_get(params[0], params[1]);
     if (!map)
         return 0;
